+++ gax-grpc/src/main/java/com/google/api/gax/grpc/nativeimage/substitutions/NettyInternalLoggerFactorySubstitutions.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Google LLC
+ * Copyright 2022 Google LLC
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
@@ -27,24 +27,45 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.google.api.gax.rpc;
+
+package com.google.api.gax.grpc.nativeimage.substitutions;
+
+import com.oracle.svm.core.annotate.Substitute;
+import com.oracle.svm.core.annotate.TargetClass;
+import io.grpc.netty.shaded.io.netty.util.internal.logging.InternalLoggerFactory;
+import io.grpc.netty.shaded.io.netty.util.internal.logging.JdkLoggerFactory;
+import java.util.function.BooleanSupplier;
 
 /**
- * Exception thrown when some resource has been exhausted, perhaps a per-user quota, or perhaps the
- * entire file system is out of space.
+ * Substitutions for {@link InternalLoggerFactory} which are needed to avoid dynamic loading of
+ * logging library.
  */
-public class ResourceExhaustedException extends ApiException {
-  public ResourceExhaustedException(Throwable cause, StatusCode statusCode, boolean retryable) {
-    super(cause, statusCode, retryable);
+@TargetClass(
+    className = "io.grpc.netty.shaded.io.netty.util.internal.logging.InternalLoggerFactory",
+    onlyWith = NettyInternalLoggerFactorySubstitutions.OnlyIfInClassPath.class)
+final class NettyInternalLoggerFactorySubstitutions {
+
+  private NettyInternalLoggerFactorySubstitutions() {}
+
+  @Substitute
+  private static InternalLoggerFactory newDefaultFactory(String name) {
+    return JdkLoggerFactory.INSTANCE;
   }
 
-  public ResourceExhaustedException(
-      String message, Throwable cause, StatusCode statusCode, boolean retryable) {
-    super(message, cause, statusCode, retryable);
-  }
+  static class OnlyIfInClassPath implements BooleanSupplier {
 
-  public ResourceExhaustedException(
-      Throwable cause, StatusCode statusCode, boolean retryable, ErrorDetails errorDetails) {
-    super(cause, statusCode, retryable, errorDetails);
+    @Override
+    public boolean getAsBoolean() {
+      try {
+        // Note: Set initialize = false to avoid initializing the class when looking it up.
+        Class.forName(
+            "io.grpc.netty.shaded.io.netty.util.internal.logging.InternalLoggerFactory",
+            false,
+            Thread.currentThread().getContextClassLoader());
+        return true;
+      } catch (ClassNotFoundException e) {
+        return false;
+      }
+    }
   }
 }