+++ gax/src/main/java/com/google/api/gax/nativeimage/substitutions/ApiClientVersionSubstitutions.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Google LLC
+ * Copyright 2022 Google LLC
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
@@ -27,24 +27,52 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.google.api.gax.rpc;
 
-/**
- * Exception thrown when some resource has been exhausted, perhaps a per-user quota, or perhaps the
- * entire file system is out of space.
- */
-public class ResourceExhaustedException extends ApiException {
-  public ResourceExhaustedException(Throwable cause, StatusCode statusCode, boolean retryable) {
-    super(cause, statusCode, retryable);
+package com.google.api.gax.nativeimage.substitutions;
+
+import com.oracle.svm.core.annotate.Alias;
+import com.oracle.svm.core.annotate.Substitute;
+import com.oracle.svm.core.annotate.TargetClass;
+import java.util.function.BooleanSupplier;
+
+/** Substitution for setting Java version correctly in the Google Java Http Client. */
+@TargetClass(
+    className =
+        "com.google.api.client.googleapis.services.AbstractGoogleClientRequest$ApiClientVersion",
+    onlyWith = ApiClientVersionSubstitutions.OnlyIfInClassPath.class)
+final class ApiClientVersionSubstitutions {
+
+  @Alias private String versionString;
+
+  @Substitute
+  public String toString() {
+    String[] tokens = versionString.split(" ");
+
+    if (tokens.length > 0 && tokens[0].startsWith("gl-java")) {
+      tokens[0] += "-graalvm";
+      return String.join(" ", tokens);
+    } else {
+      return versionString;
+    }
   }
 
-  public ResourceExhaustedException(
-      String message, Throwable cause, StatusCode statusCode, boolean retryable) {
-    super(message, cause, statusCode, retryable);
-  }
+  private ApiClientVersionSubstitutions() {}
 
-  public ResourceExhaustedException(
-      Throwable cause, StatusCode statusCode, boolean retryable, ErrorDetails errorDetails) {
-    super(cause, statusCode, retryable, errorDetails);
+  static class OnlyIfInClassPath implements BooleanSupplier {
+
+    @Override
+    public boolean getAsBoolean() {
+      try {
+        // Note: Set initialize = false to avoid initializing the class when looking it up.
+        Class.forName(
+            "com.google.api.client.googleapis.services."
+                + "AbstractGoogleClientRequest$ApiClientVersion",
+            false,
+            Thread.currentThread().getContextClassLoader());
+        return true;
+      } catch (ClassNotFoundException e) {
+        return false;
+      }
+    }
   }
 }