+++ gax/src/main/java/com/google/api/gax/rpc/UnknownException.java
@@ -44,4 +44,9 @@ public class UnknownException extends ApiException {
       String message, Throwable cause, StatusCode statusCode, boolean retryable) {
     super(message, cause, statusCode, retryable);
   }
+
+  public UnknownException(
+      Throwable cause, StatusCode statusCode, boolean retryable, ErrorDetails errorDetails) {
+    super(cause, statusCode, retryable, errorDetails);
+  }
 }